refactor(rockspec): move rocks.rockspec to its own module
