+++ rocks-lib/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod config;
 pub mod manifest;
 pub mod rocks;
+pub mod rockspec;
