+++ rocks-lib/src/rocks/mod.rs
@@ -1,13 +1,5 @@
-mod dependency;
 mod download;
-mod platform;
-mod rock_source;
-mod rockspec;
 mod unpack;
 
-pub use dependency::*;
 pub use download::*;
-pub use platform::*;
-pub use rock_source::*;
-pub use rockspec::*;
 pub use unpack::*;
